backend/requirements.txt
asgiref==3.11.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.5.5
typing_extensions==4.15.0
